Fix build failures caused by changes in cups and jinja behaviour
Changes in newer versions of jinja and cups cause builds to fail.
Since this is a long-lived ebuild, switching to a bundled version
of jinja should be preferable to patching to ensure compatibility
with newer versions of system jinja.

Patches were backported from:
https://gitweb.gentoo.org/repo/gentoo.git/commit/www-client/chromium?id=a57c78336e7c44f8d02f8591479953c181368b24
https://chromium.googlesource.com/chromium/src.git/+/29c1262dbd4f749dd514d05a571cdd809ba8cfe0%5E%21/#F0

To allow the ebuilds to build properly with >=dev-python/jinja-2.9.0
(switched to bundled version of jinja) and >=net-print/cups-2.2.0 (
patched source to include a required header file)

Issues associated with the commit:
https://github.com/elprans/atom-overlay/issues/13
https://github.com/elprans/atom-overlay/issues/14
https://bugs.funtoo.org/browse/FL-3511

Patches were applied by: Bruno Henc hencb@vorangerie.com